Fixed bug with manager. New bug found with uploader.
Updated debugs:
void Newdebugmessage(char * message, int data1, int data2, int file)
creates a debug message starting with the PID it was called on
message = string message to send (if it’s INITIT, It will initialize a file with a log header)
data1 = int to put in file. -42 won’t put a number
data2 = int to put in file. -42 won’t put a number
file = int to determine which debug file to put it in (debug#.txt). -42 will set the filename to debug.txt